Initial import for MineMe
A Powershell script that monitors the number of seconds a particular
window is active and then logs data to a file. logs to csv with the
window title and the total number of seconds presumed to be focussed.
Polls every second. Does not log if it detects an idle time of more than
60 seconds.